Remove the _toolchain_versions ("dead") code.
As the main purpose of this code was to generate the report section of
toolchain modules and this is already done by Common::Client.

The problem, is that as a side effect it populated $result->{toolchain},
that was used to determine if it should call _parse_tap_harness or
_parse_test_harness. So it required a little more effort to mantain
everything else working without messing too much with the code.